requirements.txt
fastapi
uvicorn
requests
numpy
python-multipart
databricks-sql-connector
python-dotenv